requirements.txt
python-dotenv
spotipy
pyspark
kafka
